+++ app/src/main/java/de/kindermaenner/monatsblitz/domain/model/MatchKey.kt
@@ -1,3 +1,3 @@
 package de.kindermaenner.monatsblitz.domain.model
 
-data class MatchKey(val tournamentId : Int, val player1Id: Int, val player2Id: Int, val leg: Leg)
+data class MatchKey(val tournamentId : Long, val player1Id: Long, val player2Id: Long, val leg: Leg)
