+++ app/src/main/java/de/kindermaenner/monatsblitz/domain/model/NewGame.kt
@@ -1,3 +1,3 @@
 package de.kindermaenner.monatsblitz.domain.model
 
-data class NewGame(val tournamentId : Int, val leg: Leg, val player1Id: Int, val player2Id: Int)
+data class NewGame(val tournamentId : Long, val leg: Leg, val player1Id: Long, val player2Id: Long)
