+++ app/src/main/java/de/kindermaenner/monatsblitz/domain/model/NewTournament.kt
@@ -4,5 +4,5 @@ import java.time.LocalDate
 
 data class NewTournament( val Mode : GameMode,
                           val Date : LocalDate,
-                          val playerIds: List<Int>,
+                          val playerIds: List<Long>,
                           val doubleRound: Boolean)
