+++ app/src/main/java/de/kindermaenner/monatsblitz/domain/model/Player.kt
@@ -1,6 +1,6 @@
 package de.kindermaenner.monatsblitz.domain.model
 
-data class Player(val id: Int, val Name : String, val Vorname : String) {
+data class Player(val id: Long, val Name : String, val Vorname : String) {
     val fullName: String
         get() = "$Vorname $Name"
 }
