+++ app/src/main/java/de/kindermaenner/monatsblitz/infrastructure/persistence/room/entity/TournamentPlayerCrossRef.kt
@@ -7,7 +7,7 @@ import androidx.room.Entity
     primaryKeys = ["tournamentId", "playerId"]
 )
 data class TournamentPlayerCrossRef(
-    val tournamentId: Int,
+    val tournamentId: Long,
 
-    val playerId: Int
+    val playerId: Long
 )
